requirements.txt
fastapi==0.116.1
pytest-asyncio==0.24.0
pydantic>=2.11.7
pydantic[email]>=2.11.7
passlib==1.7.4
python-jose[cryptography]==3.5.0
bcrypt==3.2.0
uvicorn==0.35.0
httpx==0.28.1
python-dotenv==1.1.1
pytest==8.4.1
motor==3.7.1
fuzzywuzzy==0.18.0
bson==0.5.10
pydantic-settings==2.10.1
pymongo==4.10.0
orjson==3.11.1